# unit: use more efficient sort routine