@jbduncan 2024-04-07
These benchmark results seem to prove that there are performance improvements:

```
goos: darwin
goarch: arm64
pkg: gonum.org/v1/gonum/unit
                             │ unit/benchmarks_dimensions_string_old.txt │ unit/benchmarks_dimensions_string_new.txt │
                             │                  sec/op                   │      sec/op        vs base                │
DimensionsString/oneDim-8                                    175.8n ± 0%         147.6n ± 0%  -16.07% (p=0.000 n=10)
DimensionsString/threeDims-8                                 488.7n ± 0%         455.1n ± 0%   -6.88% (p=0.000 n=10)
DimensionsString/allDims-8                                   1.457µ ± 0%         1.417µ ± 1%   -2.75% (p=0.000 n=10)
geomean                                                      500.3n              456.5n        -8.74%
```