@jbduncan 2024-04-14
> Please separate the benchmark addition into a separate commit, prior to the logic change so that the perf impact can be confirm eternally.

Alrighty! Done.

I'm ready for another review whenever you are.